Changed calculation logic to use BigDecimal and created BaseOperation abstract class

+++ CalculatorServer/app/src/main/java/com/example/calculatorserver/operations/OperationInterface.java
@@ -1,5 +1,7 @@
 package com.example.calculatorserver.operations;
 
+import java.math.BigDecimal;
+
 public interface OperationInterface {
-    double performOperation(double num1, double num2);
+    BigDecimal performOperation(double num1, double num2);
 }
